Imagen/requirements.txt
accelerate==0.18.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beartype==0.13.1
blinker==1.6.2
blis==0.7.9
braceexpand==0.1.7
brotlipy==0.7.0
cachetools==5.3.0
catalogue==2.0.8
cchardet==2.1.7
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
clip==0.2.0
cmake==3.26.3
confection==0.0.4
contexttimer==0.3.3
contourpy==1.0.7
cryptography @ file:///croot/cryptography_1677533068310/work
cycler==0.10.0
cymem==2.0.7
datasets==2.11.0
decorator==5.1.1
decord==0.6.0
dill==0.3.6
distlib==0.3.6
docker-pycreds==0.4.0
einops==0.6.0
einops-exts==0.0.4
ema-pytorch==0.2.2
entrypoints==0.4
executing==1.2.0
fairscale==0.4.4
filelock==3.11.0
flit_core @ file:///croot/flit-core_1679397103445/work/source/flit_core
fonttools==4.22.0
frozenlist==1.3.3
fsspec==2023.4.0
ftfy==6.1.1
gitdb==4.0.10
GitPython==3.1.31
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645438755360/work
huggingface-hub==0.13.4
identify==2.5.24
idna @ file:///croot/idna_1666125576474/work
imageio==2.28.1
imagen-pytorch==1.24.2
importlib-metadata==6.6.0
importlib-resources==5.12.0
iopath==0.1.10
ipython==8.13.2
jedi==0.18.2
Jinja2 @ file:///croot/jinja2_1666908132255/work
joblib==1.2.0
jsonschema==4.17.3
kaggle==1.5.13
kiwisolver==1.0.1
kornia==0.6.11
langcodes==3.3.0
lazy_loader==0.2
lightning-utilities==0.8.0
lion-pytorch==0.0.7
lit==16.0.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
murmurhash==1.0.9
networkx==3.1
nodeenv==1.7.0
numpy==1.24.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
openai==0.27.6
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
packaging==23.1
pandas==2.0.0
parso==0.8.3
pathtools==0.1.2
pathy==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.0
plotly==5.14.1
portalocker==2.7.0
pre-commit==3.3.1
preshed==3.0.8
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pycocoevalcap==1.2
pycocotools==2.0.6
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.10.7
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.8.2
python-magic==0.4.27
python-slugify==8.0.1
pytorch-fid==0.3.0
pytorch-lightning==2.0.1.post0
pytorch-warmup==0.1.1
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
responses==0.18.0
rich==13.3.5
salesforce-lavis==1.0.2
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
seaborn==0.12.2
sentencepiece==0.1.98
sentry-sdk==1.21.1
setproctitle==1.3.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
smart-open==6.3.0
smmap==5.0.0
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
stack-data==0.6.2
streamlit==1.22.0
sympy @ file:///croot/sympy_1668202399572/work
tenacity==8.2.2
text-unidecode==1.3
thinc==8.1.10
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.4.12
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.0
torchaudio==2.0.0
torchmetrics==0.11.4
torchvision==0.15.1
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
transformers==4.26.1
triton==2.0.0
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3 @ file:///croot/urllib3_1680254681959/work
validators==0.20.0
virtualenv==20.23.0
wandb==0.15.0
wasabi==1.1.1
watchdog==3.0.0
wcwidth==0.2.6
webdataset==0.2.48
xgboost==1.7.5
xxhash==3.2.0
yarl==1.8.2
zipp==3.15.0
